requirements.txt
pillow==10.0.0
pynput==1.7.6
psutil==5.9.5
pygetwindow==0.0.9 